Moved tests from SequenceCompareTo.String.cs to SequenceCompareTo.T.cs
